rename algo and dir
